packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexum-sdk"
version = "0.1.0"
description = "Python SDK for Nexum — Durable Execution Engine for LLM Agents"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "pydantic>=2.0.0",
    "pydantic-core>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
